Support Node 8 & prevent username cache conflicts
Now builds and at least starts on Node.js v8, although not extensively
tested on that version.

Feat: Chrome profile directory names (for caching, etc.) now also
include the user username in order to prevent an older login from being
cached when a different user is intended to be logged in.